refactor default openai parameters in ChatWrapper
